# Improve Array#unshift performance on v8 > 7.1