--- a/opal/corelib/array.rb
+++ b/opal/corelib/array.rb
@@ -33,6 +33,17 @@ class Array < `Array`
       return r;
     }
 
+    function unshiftOne(list, elem) {
+      var len = list.length - 1;
+      list.push(list[len]);
+      while (len) {
+        list[len] = list[len - 1];
+        len--;
+      }
+      list[0] = elem;
+      return list.length;
+    }
+
     function toArraySubclass(obj, klass) {
       if (klass.$$name === Opal.Array) {
         return obj;
@@ -2292,7 +2303,7 @@ class Array < `Array`
   def unshift(*objects)
     %x{
       for (var i = objects.length - 1; i >= 0; i--) {
-        self.unshift(objects[i]);
+        unshiftOne(self, objects[i]);
       }
     }