### Command

```
$ bundle exec rake bench:ips FILE=benchmark-ips/bm_array_unshift.rb
```

### Before

#### Node 10

```
=== Files: benchmark-ips/bm_array_unshift.rb
bundle exec opal --dynamic-require ignore --missing-require ignore -ropal/platform -gbenchmark-ips -rbenchmark/ips -A benchmark-ips/bm_array_unshift.rb
Warming up --------------------------------------
             unshift    23.351k i/100ms
Calculating -------------------------------------
             unshift    373.961k (± 2.3%) i/s -      1.891M in   5.060583s
```

#### Node 14

```
=== Files: benchmark-ips/bm_array_unshift.rb
bundle exec opal --dynamic-require ignore --missing-require ignore -ropal/platform -gbenchmark-ips -rbenchmark/ips -A benchmark-ips/bm_array_unshift.rb
Warming up --------------------------------------
             unshift   698.000  i/100ms
Calculating -------------------------------------
             unshift      7.188k (± 0.7%) i/s -     36.296k in   5.049684s
```

### After

#### Node 10

```
=== Files: benchmark-ips/bm_array_unshift.rb
bundle exec opal --dynamic-require ignore --missing-require ignore -ropal/platform -gbenchmark-ips -rbenchmark/ips -A benchmark-ips/bm_array_unshift.rb
Warming up --------------------------------------
             unshift    12.904k i/100ms
Calculating -------------------------------------
             unshift    187.988k (± 1.8%) i/s -    941.992k in   5.012631s
```

#### Node 14

```
=== Files: benchmark-ips/bm_array_unshift.rb
bundle exec opal --dynamic-require ignore --missing-require ignore -ropal/platform -gbenchmark-ips -rbenchmark/ips -A benchmark-ips/bm_array_unshift.rb
Warming up --------------------------------------
             unshift    19.317k i/100ms
Calculating -------------------------------------
             unshift    247.057k (± 1.7%) i/s -      1.236M in   5.005546s
```

### Summary

With this change, the code is now 1.98x times slower when using Node 10 but 34.37x times faster when using Node 14.

We cannot beat the highly optimized code produced by v8 pre 7.1 (Node 10) but the code is now "only" 1.5x times slower when using a recent version of v8/Node.