@Mogztter 2020-10-01
@elia Your code is more elegant but slightly slower on several `list` lengths and `elems` lengths.
The best-case scenario for your implementation is when the `list` is small (10 elements) and there are a few elements to unshift (10 elements).

Here's a comparison with a few variations on the `list` lengths and `elems` lengths.


The following variation is the best-case scenario for your implementation:

```
list.length = 10
elems.length = 10
```

| Round | Array#unshift (base) | unshiftOne | unshiftAll (elia)  | unshiftAll (mogztter) |
| -- | -- | -- | -- | -- |
| 1 | 115,111 ops/sec<br/> ±0.55%<br/> (95 runs sampled) | 1,780,182 ops/sec<br/> ±2.44%<br/> (87 runs sampled) | 3,654,839 ops/sec :star: <br/> ±1.56% <br/>(92 runs sampled)  | 2,593,891 ops/sec<br/> ±1.16%<br/> (91 runs sampled)
| 2 | 111,308 ops/sec <br/>±1.69%<br/> (88 runs sampled) | 1,932,104 ops/sec<br/> ±0.51%<br/> (96 runs sampled) | 3,662,971 ops/sec :star: <br/> ±1.50% <br/>(88 runs sampled) | 2,695,324 ops/sec<br/> ±0.49% <br/>(92 runs sampled)
| 3 | 116,501 ops/sec<br/> ±0.49% <br/>(93 runs sampled) | 1,913,965 ops/sec<br/> ±0.37% <br/>(92 runs sampled) | 3,750,125 ops/sec :star: <br/>±0.68% <br/>(93 runs sampled) | 2,620,791 ops/sec <br/>±0.69% <br/>(90 runs sampled)


```
list.length = 100
elems.length = 10
```

| Round | Array#unshift (base) | unshiftOne | unshiftAll (elia)  | unshiftAll (mogztter) |
| -- | -- | -- | -- | -- |
| 1 | 19,832 ops/sec <br/>±0.63% <br/>(96 runs sampled) | 632,687 ops/sec<br/> ±0.38% <br/>(96 runs sampled) | 687,256 ops/sec <br/>±0.45% <br/>(96 runs sampled) | 1,611,607 ops/sec :star:<br/> ±0.25%<br/> (97 runs sampled)
| 2 | 19,258 ops/sec <br/>±0.27% <br/>(97 runs sampled) | 610,101 ops/sec <br/>±0.32% <br/>(92 runs sampled) | 673,388 ops/sec <br/>±0.26% <br/>(93 runs sampled) | 1,571,381 ops/sec :star: <br/>±0.43% <br/>(93 runs sampled)
| 3 | 19,326 ops/sec <br/>±0.40% <br/>(92 runs sampled) | 621,284 ops/sec<br/> ±0.28% <br/>(91 runs sampled) | 672,708 ops/sec <br/>±0.32% <br/>(93 runs sampled)| 1,570,496 ops/sec :star: <br/> ±0.42%<br/> (93 runs sampled)

```
list.length = 100
elems.length = 1
```

| Round | Array#unshift (base) | unshiftOne | unshiftAll (elia)  | unshiftAll (mogztter) |
| -- | -- | -- | -- | -- |
| 1 | 187,987 ops/sec  <br/>±0.40% <br/> (95 runs sampled) | 3,031,309 ops/sec :star:  <br/> ±0.51% <br/> (92 runs sampled) | 863,585 ops/sec <br/> ±0.92% <br/> (92 runs sampled) | 3,027,611 ops/sec  <br/>±0.38% <br/> (92 runs sampled)
| 2 | 184,408 ops/sec <br/> ±0.36% <br/> (93 runs sampled) | 2,968,972 ops/sec  <br/>±0.80% <br/> (94 runs sampled) | 883,129 ops/sec  <br/>±0.39% <br/> (96 runs sampled) | 3,013,784 ops/sec :star:  <br/> ±0.37% <br/> (92 runs sampled)
| 3 | 186,859 ops/sec  <br/>±0.45% <br/>(93 runs sampled) | 3,028,500 ops/sec :star:  <br/> ±0.45% <br/> (91 runs sampled) | 894,202 ops/sec <br/> ±0.40%  <br/>(94 runs sampled) | 2,954,361 ops/sec  <br/>±0.41%  <br/>(91 runs sampled)


```
list.length = 1001
elems.length = 10
```

| Round | Array#unshift (base) | unshiftOne | unshiftAll (elia)  | unshiftAll (mogztter) |
| -- | -- | -- | -- | -- |
| 1 | 2,076 ops/sec <br/>±0.40% <br/> (95 runs sampled) | 91,423 ops/sec  <br/>±0.37% <br/> (95 runs sampled) | 73,452 ops/sec <br/> ±0.30%  <br/>(94 runs sampled) | 461,994 ops/sec :star: <br/> ±0.48%  <br/>(92 runs sampled)
| 2 | 2,082 ops/sec  <br/>±0.35%  <br/>(94 runs sampled) | 92,391 ops/sec  <br/>±0.24% <br/> (96 runs sampled) | 76,563 ops/sec  <br/>±0.18%  <br/>(96 runs sampled) | 472,779 ops/sec :star: <br/> ±0.29% <br/> (92 runs sampled)
| 3 | 2,064 ops/sec  <br/>±0.51%  <br/>(92 runs sampled) | 91,019 ops/sec  <br/>±0.58% <br/> (91 runs sampled) | 75,691 ops/sec  <br/>±0.33%  <br/>(94 runs sampled) | 464,082 ops/sec :star:  <br/>±0.44% <br/>(95 runs sampled)

I don't explain why my implementation is significantly faster in this last case.
Maybe my code is hitting an optimization.