@Mogztter 2021-07-03
Thanks for the merge @hmdne.
I get similar performance with the latest changes (using Node v14.15.4):

```
arrayUnshift {array.length: 1001, elem.length: 10} x 1,851 ops/sec ±2.36% (86 runs sampled)
unshiftOne {array.length: 1001, elem.length: 1} x 78,123 ops/sec ±1.91% (88 runs sampled)
unshiftAllElia {array.length: 1001, elem.length: 10} x 62,608 ops/sec ±2.55% (87 runs sampled)
unshiftAllHmdne {array.length: 1001, elem.length: 10} x 376,396 ops/sec ±2.76% (84 runs sampled)
unshiftAllGuillaume {array.length: 1001, elem.length: 10} x 385,477 ops/sec ±1.87% (89 runs sampled)
```

Do you see any improvements on your side?