@hmdne 2021-07-08
My version is basically your `unshiftAllGuillaume` with a shortpath if an empty array is passed. I should have noted that though. I tried benchmarking some hybrid approach with either `unshiftOne` or `unshiftAllGuillaume`, but the gains from the first one were negligible (as shown by your benchmark).